// test/integration/test/ports/_fixtures/naming-conflict-builtin-vs-model/generated/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
type DefaultLiteralValue<CodecId extends string, _Encoded> = CodecId extends keyof CodecTypes
  ? CodecTypes[CodecId]['output']
  : _Encoded;

export type FieldOutputTypes = {
  readonly public: {
    readonly Keys: {
      readonly id: Char<24>;
      readonly isUserProvidedType: CodecTypes['pg/bool@1']['output'];
    };
    readonly Promise: {
      readonly id: Char<24>;
      readonly isUserProvidedType: CodecTypes['pg/bool@1']['output'];
    };
    readonly RelationHolderKeys: {
      readonly id: Char<24>;
      readonly modelId: CodecTypes['pg/text@1']['output'];
    };
    readonly RelationHolderPromise: {
      readonly id: Char<24>;
      readonly modelId: CodecTypes['pg/text@1']['output'];
    };
    readonly RelationHolderResult: {
      readonly id: Char<24>;
      readonly modelId: CodecTypes['pg/text@1']['output'];
    };
    readonly RelationHolderUnion: {
      readonly id: Char<24>;
      readonly modelId: CodecTypes['pg/text@1']['output'];
    };
    readonly Result: {
      readonly id: Char<24>;
      readonly isUserProvidedType: CodecTypes['pg/bool@1']['output'];
    };
    readonly Union: {
      readonly id: Char<24>;
      readonly isUserProvidedType: CodecTypes['pg/bool@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Keys: {
      readonly id: CodecTypes['sql/char@1']['input'];
      readonly isUserProvidedType: CodecTypes['pg/bool@1']['input'];
    };
    readonly Promise: {
      readonly id: CodecTypes['sql/char@1']['input'];
      readonly isUserProvidedType: CodecTypes['pg/bool@1']['input'];
    };
    readonly RelationHolderKeys: {
      readonly id: CodecTypes['sql/char@1']['input'];
      readonly modelId: CodecTypes['pg/text@1']['input'];
    };
    readonly RelationHolderPromise: {
      readonly id: CodecTypes['sql/char@1']['input'];
      readonly modelId: CodecTypes['pg/text@1']['input'];
    };
    readonly RelationHolderResult: {
      readonly id: CodecTypes['sql/char@1']['input'];
      readonly modelId: CodecTypes['pg/text@1']['input'];
    };
    readonly RelationHolderUnion: {
      readonly id: CodecTypes['sql/char@1']['input'];
      readonly modelId: CodecTypes['pg/text@1']['input'];
    };
    readonly Result: {
      readonly id: CodecTypes['sql/char@1']['input'];
      readonly isUserProvidedType: CodecTypes['pg/bool@1']['input'];
    };
    readonly Union: {
      readonly id: CodecTypes['sql/char@1']['input'];
      readonly isUserProvidedType: CodecTypes['pg/bool@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly keys: {
      readonly id: Char<24>;
      readonly isUserProvidedType: CodecTypes['pg/bool@1']['output'];
    };
    readonly promise: {
      readonly id: Char<24>;
      readonly isUserProvidedType: CodecTypes['pg/bool@1']['output'];
    };
    readonly relation_holder_keys: {
      readonly id: Char<24>;
      readonly modelId: CodecTypes['pg/text@1']['output'];
    };
    readonly relation_holder_promise: {
      readonly id: Char<24>;
      readonly modelId: CodecTypes['pg/text@1']['output'];
    };
    readonly relation_holder_result: {
      readonly id: Char<24>;
      readonly modelId: CodecTypes['pg/text@1']['output'];
    };
    readonly relation_holder_union: {
      readonly id: Char<24>;
      readonly modelId: CodecTypes['pg/text@1']['output'];
    };
    readonly result: {
      readonly id: Char<24>;
      readonly isUserProvidedType: CodecTypes['pg/bool@1']['output'];
    };
    readonly union: {
      readonly id: Char<24>;
      readonly isUserProvidedType: CodecTypes['pg/bool@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly keys: {
      readonly id: CodecTypes['sql/char@1']['input'];
      readonly isUserProvidedType: CodecTypes['pg/bool@1']['input'];
    };
    readonly promise: {
      readonly id: CodecTypes['sql/char@1']['input'];
      readonly isUserProvidedType: CodecTypes['pg/bool@1']['input'];
    };
    readonly relation_holder_keys: {
      readonly id: CodecTypes['sql/char@1']['input'];
      readonly modelId: CodecTypes['pg/text@1']['input'];
    };
    readonly relation_holder_promise: {
      readonly id: CodecTypes['sql/char@1']['input'];
      readonly modelId: CodecTypes['pg/text@1']['input'];
    };
    readonly relation_holder_result: {
      readonly id: CodecTypes['sql/char@1']['input'];
      readonly modelId: CodecTypes['pg/text@1']['input'];
    };
    readonly relation_holder_union: {
      readonly id: CodecTypes['sql/char@1']['input'];
      readonly modelId: CodecTypes['pg/text@1']['input'];
    };
    readonly result: {
      readonly id: CodecTypes['sql/char@1']['input'];
      readonly isUserProvidedType: CodecTypes['pg/bool@1']['input'];
    };
    readonly union: {
      readonly id: CodecTypes['sql/char@1']['input'];
      readonly isUserProvidedType: CodecTypes['pg/bool@1']['input'];
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;